+++ lib/src/form_builder/date_time_textfield_parser.dart
@@ -88,7 +88,7 @@ class _DateTimeTextFieldCreatorState extends State<DateTimeTextFieldCreator> {
             : true;
         if (!isVisible) widget.controller.text = "";
         return (!isVisible)
-            ? Container()
+            ? SizedBox.shrink()
             : Neumorphic(
                 child: Container(
                   width: (size.width * (1 / (widget.map.total + 0.5))),