+++ lib/src/form_builder/email_textfield_parser.dart
@@ -103,7 +103,7 @@ class _EmailTextFieldCreatorState extends State<EmailTextFieldCreator> {
         }
         if (!isVisible) widget.controller.text = "";
         return (!isVisible)
-            ? Container()
+            ? SizedBox.shrink()
             : Neumorphic(
                 margin: EdgeInsets.symmetric(horizontal: 16, vertical: 12),
                 padding: EdgeInsets.symmetric(horizontal: 12, vertical: 6),