+++ lib/src/form_builder/phone_textfield_parser.dart
@@ -104,7 +104,7 @@ class _PhoneTextFieldCreatorState extends State<PhoneTextFieldCreator> {
         }
         if (!isVisible) widget.controller.text = "";
         return (!isVisible)
-            ? Container()
+            ? SizedBox.shrink()
             : Neumorphic(
                 margin: EdgeInsets.symmetric(horizontal: 16, vertical: 12),
                 padding: EdgeInsets.symmetric(horizontal: 12, vertical: 6),