+++ lib/src/form_builder/url_textfield_parser.dart
@@ -104,7 +104,7 @@ class _UrlTextFieldCreatorState extends State<UrlTextFieldCreator> {
         }
         if (!isVisible) widget.controller.text = "";
         return (!isVisible)
-            ? Container()
+            ? SizedBox.shrink()
             : Neumorphic(
                 margin: EdgeInsets.symmetric(horizontal: 16, vertical: 12),
                 padding: EdgeInsets.symmetric(horizontal: 12, vertical: 6),