+++ lib/src/models/form_collection.dart
@@ -29,7 +29,7 @@ class FormCollection {
   /// Returns a [FormCollection] from a [Map<String, dynamic>].
   factory FormCollection.fromJson(Map<String, dynamic> json) => FormCollection(
         display: json["display"],
-        components:  List<Component>.from(
+        components: List<Component>.from(
             json["components"].map((x) => Component.fromJson(x))),
       );
 
@@ -210,10 +210,11 @@ class Component {
 
   /// Returns a [Component] from a [String] that possees a json.
   factory Component.fromJson(Map<String, dynamic> json) => Component(
-        component: (json["components"] == null || !json.containsKey('components'))
-            ? null
-            : List<Component>.from(
-                json["components"].map((x) => Component.fromJson(x))),
+        component:
+            (json["components"] == null || !json.containsKey('components'))
+                ? null
+                : List<Component>.from(
+                    json["components"].map((x) => Component.fromJson(x))),
         columns: (json["columns"] == null || !json.containsKey('columns'))
             ? null
             : List<Component>.from(
@@ -231,18 +232,21 @@ class Component {
             (json["showWordCount"] == null) ? null : json["showWordCount"],
         decimalLimit:
             json.containsKey('decimalLimit') ? json['decimalLimit'] ?? 4 : 4,
-        penColor: (json["penColor"] == null) ? "black" : json["penColor"],
+        penColor: json.containsKey("penColor")
+            ? json["penColor"].toString().toLowerCase() ?? "black"
+            : "black",
         footer: (json["footer"] == null) ? null : json["footer"],
         backgroundColor: (json["backgroundColor"] == null)
             ? "rgb(255, 255, 255)"
             : json["backgroundColor"],
         mask: (json["mask"] == null) ? false : json["mask"],
         data: json["data"] == null ? null : Data.fromJson(json["data"]),
-        defaultValue: json["defaultValue"] == null  || !json.containsKey('defaultValue')
-            ? null
-            : (json["defaulValue"] is List<String>)
-                ? List<String>.from(json["defaultValue"].map((x) => x))
-                : json["defaultValue"],
+        defaultValue:
+            json["defaultValue"] == null || !json.containsKey('defaultValue')
+                ? null
+                : (json["defaulValue"] is List<String>)
+                    ? List<String>.from(json["defaultValue"].map((x) => x))
+                    : json["defaultValue"],
         labelPosition:
             json["labelPosition"] == null ? null : json["labelPosition"],
         prefix: json["prefix"] == null ? null : json["prefix"],
@@ -267,7 +271,7 @@ class Component {
         rightIcon: json["rightIcon"] == null ? null : json["rightIcon"],
         description: json["description"] == null ? null : json["description"],
         tooltip: json["tooltip"] == null ? null : json["tooltip"],
-        tags: json["tags"] == null  || !json.containsKey('tags')
+        tags: json["tags"] == null || !json.containsKey('tags')
             ? null
             : List<String>.from(json["tags"].map((x) => x)),
       );
@@ -368,7 +372,9 @@ class Data {
 
   /// Returns a [Data] from a [String] that possees a json.
   factory Data.fromJson(Map<String, dynamic> json) => Data(
-        values: json.containsKey('values') ? List<Value>.from(json["values"].map((x) => Value.fromJson(x))) : [],
+        values: json.containsKey('values')
+            ? List<Value>.from(json["values"].map((x) => Value.fromJson(x)))
+            : [],
         url: json["url"] == null ? "" : json["url"],
       );